refactor: extract ImageProcessor and VideoImporter into separate modules
- Move ImageProcessor and VideoImporter classes from Image2D.h to dedicated headers
- Create ImageProcessor.h and VideoImporter.h with corresponding .cpp implementations
- Add core header wrappers under ncorr/core/ directory
- Update CMakeLists.txt to include new source files
- Add .DS_Store to .gitignore
- Reduce Image2D.h coupling by splitting image processing and video import concerns
