+++ include/ncorr/core/image_processor.hpp
@@ -1,0 +1,6 @@
+#ifndef NCORR_CORE_IMAGE_PROCESSOR_HPP
+#define NCORR_CORE_IMAGE_PROCESSOR_HPP
+
+#include "../../ImageProcessor.h"
+
+#endif
