+++ include/ncorr/core/video_importer.hpp
@@ -1,0 +1,6 @@
+#ifndef NCORR_CORE_VIDEO_IMPORTER_HPP
+#define NCORR_CORE_VIDEO_IMPORTER_HPP
+
+#include "../../VideoImporter.h"
+
+#endif
